:recycle: refactor: make enums package, change enum class name(#57)
