+++ src/main/java/com/praise/push/common/error/enums/PraiseUpErrorCode.java
@@ -1,11 +1,11 @@
-package com.praise.push.common;
+package com.praise.push.common.error.enums;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum CustomErrorCode {
+public enum PraiseUpErrorCode {
     CUSTOM_ERROR("custom error message 작성");
 
     private final String statusMessage;
